pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miragepot"
version = "0.1.0"
description = "AI-Driven Adaptive SSH Honeypot using local LLM"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Evin Brijesh", email = "evinbrijesh@example.com"}
]
maintainers = [
    {name = "Evin Brijesh", email = "evinbrijesh@example.com"}
]
keywords = [
    "honeypot",
    "ssh",
    "security",
    "ai",
    "llm",
    "cybersecurity",
    "deception",
    "threat-detection"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Networking :: Monitoring",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "paramiko>=3.0.0",
    "ollama>=0.1.0",
    "colorama>=0.4.6",
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
miragepot = "miragepot.__main__:main"

[project.urls]
Homepage = "https://github.com/evinbrijesh/MiragePot"
Documentation = "https://github.com/evinbrijesh/MiragePot#readme"
Repository = "https://github.com/evinbrijesh/MiragePot.git"
Issues = "https://github.com/evinbrijesh/MiragePot/issues"
Changelog = "https://github.com/evinbrijesh/MiragePot/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["miragepot*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
miragepot = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["miragepot"]
branch = true
omit = [
    "miragepot/__main__.py",
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG001"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
]
